// appBelotur/src/app/models/formulario.ts
export class Formulario {
  id: any;
  nome: string;
  rg: string;
  cpf: any;
  sexo: string;
  nacionalidade: string;
  nomeMae: string;
  nomePai: string;
  cep: any;
  cidade: string;
  bairro: string;
  endereco: string;
  email: string;
  telefone: string;
  celular: string;
  foto: any;
  antecedenteCriminal: any;
  formularioAdesao: any;
  comprovanteEndereco: any;
  documentoFrente: any;
  documentoVerso: any;
  permissao: any;
  senha: any;
}